Move updatePublishingStatus to repository model
